Add files via upload
A Wikipedia-like site that you can use to store and view random entries. The purpose of this project was to get better with creating websites and django framework.

Used languages: Python, CSS, HTML, JavaScript.

If you want to turn it on, just download the file and turn it on using django commands!