Add files via upload
A selling app like Amazon, that you can list, comments and bid items on. The purpose of creating this project was to get better with django web framework and also to get some sense with OOP and SQL migrations.

Used languages: Python, CSS, HTML, SQL, Django framework.

If you want to turn it on, just download the file and turn it on using django commands!