Add files via upload
A web app for e-mail management. The purpose of this project was to get better at OOP, web programing with Django and also to boost-up JavaScript skills.

Used languages: JavaScript, Python, CSS, HTML.

If you want to turn it on, just download the file and turn it on using django commands!